+++ data/db.json
@@ -5,7 +5,7 @@
          "avatarImg":"../src/assets/img/1702865313114.jpeg",
          "nameUser":"Alan Ferreira",
          "curso": "Desenvolvimento de sistemas",
-         "description":"Salve pessoal, Alguem pode me ajudar com o codigo abaixo? ta dando muito problema",
+         "description":"Salve pessoal, Alguem pode me ajudar com o codigo abaixo? ta dando muito problema Salve pessoal",
          "img": "../src/assets/img/problemaCodigo.png"
       },
       {
