+++ src/components/Post.tsx
@@ -12,9 +12,20 @@ import {
   Image,
   Box,
   Heading,
+  Menu,
+  MenuButton,
+  MenuList,
+  MenuItem,
 } from "@chakra-ui/react";
-import { ThumbsUp, MessageSquare } from "lucide-react";
+import {
+  ThumbsUp,
+  MessageSquare,
+  Delete,
+  Ellipsis,
+  Trash2,
+} from "lucide-react";
 import { Task } from "../types/types";
+import { Menu as MenuLogo } from "lucide-react";
 
 export default function Post() {
   const [tasks, setTasks] = useState<Task[]>([]);
@@ -36,34 +47,45 @@ export default function Post() {
   return (
     <>
       {tasks.map((task) => (
-        <Card key={task.id} marginLeft={""} w={"full"} padding={"5px"}>
+        <Card
+          key={task.id}
+          marginLeft={""}
+          w={"full"}
+          alignItems={"left"}
+          padding={"0"}
+        >
           <CardHeader w={"full"} fontSize={"sm"}>
             <Flex>
-              <Flex flex="1" gap="1" alignItems="center" flexWrap="wrap">
-                <Avatar
-                  name="Segun Adebayo"
-                  src={task.avatarImg}
-                  w={"60px"}
-                  h={"60px"}
-                />
+              <Flex flex="1" gap="1" alignItems="stretch">
+                <Avatar src={task.avatarImg} w={"60px"} h={"60px"} />
                 <Box>
                   <Heading size="sm">{task.nameUser}</Heading>
-                  <Text>{task.curso}</Text>
+                  <Text fontSize={"small"}>{task.curso}</Text>
                 </Box>
               </Flex>
-              <IconButton
-                variant="ghost"
-                colorScheme="gray"
-                aria-label="See menu"
-              />
+
+              <Menu>
+                <MenuButton
+                  as={IconButton}
+                  aria-label="Options"
+                  icon={<Ellipsis />}
+                  variant="outline"
+                />
+                <MenuList>
+                  <MenuItem icon={<Trash2 />}>Excluir Postagem</MenuItem>
+                </MenuList>
+              </Menu>
             </Flex>
           </CardHeader>
-          <CardBody maxH={"120px"} fontSize={"small"}>
-            <Text>{task.description}</Text>
+          <CardBody maxH={"120px"} maxW={"600px"} fontSize={"small"} pt={0}>
+            <Text fontSize={{ sm: "10px", md: "15px" }}>
+              {task.description}
+            </Text>
           </CardBody>
 
-          <div className="flex justify-center">
+          <div className="flex justify-center m-0">
             <Image
+              flexWrap={"wrap"}
               objectFit="cover"
               src={task.img}
               padding={"5px"}
